+++ src/filegrail/sources/__init__.py
@@ -11,6 +11,7 @@ from .quarantine import collect_quarantine_events, read_quarantine
 from .recent import collect_recent_files
 from .shell import collect_shell_history
 from .shortcut import collect_windows_recent, read_shortcuts
+from .sidecar import read_sidecar
 from .xmp import read_xmp
 
 __all__ = [
@@ -27,6 +28,7 @@ __all__ = [
     "read_quarantine",
     "collect_windows_recent",
     "read_shortcuts",
+    "read_sidecar",
     "is_archive",
     "list_members",
     "inherited_origin",